Developing Day 2
Changed the decimal to double as of course it made more sense, and doubles can go to 10e300 so then yeah. also started on the ui work a bit more and then I guess that was kind of it, also made theta working.

+++ Assets/Scripts/VariableClass.cs
@@ -3,11 +3,11 @@ using System.Numerics;
 public class VariableClass
 {
     public string name { get; }
-    public decimal value { get; set; }
-    public decimal initialValue { get; set; } //In case we need to reset or something idk.
+    public double value { get; set; }
+    public double initialValue { get; set; } //In case we need to reset or something idk.
     public int level { get; set; }
 
-    public VariableClass(string name, decimal value) //Regular initializer.
+    public VariableClass(string name, double value) //Regular initializer.
     {
         this.name = name;
         this.value = value;
